fix(remotion): break circular import causing SplitScene TDZ runtime error
components.ts imported SplitScene which imported componentsById back from
components.ts; meta.ts also imported LAYOUT_SLOT_COUNTS from SplitScene.tsx
during module init while SplitScene.tsx was still evaluating, producing
"Cannot access 'SplitScene' before initialization".

- Extract base components map into componentsBase.ts (all components except
  SplitScene). SplitScene.tsx now imports componentsByIdBase from there
  instead of components.ts, eliminating that cycle.
- Move SplitLayout type and LAYOUT_SLOT_COUNTS into a new layout.ts so meta.ts
  no longer needs to re-enter SplitScene.tsx during init.
- components.ts simply spreads componentsByIdBase and adds SplitScene.
